+++ ios/Classes/SwiftConnectycubeFlutterCallKitPlugin.swift
@@ -96,10 +96,8 @@ public class SwiftConnectycubeFlutterCallKitPlugin: NSObject, FlutterPlugin {
         }
         else if(call.method == "reportCallAccepted"){
             let callId = arguments["session_id"] as! String
-            let callType = arguments["call_type"] as! Int
-            let videoEnabled = callType == 1
-            
-            SwiftConnectycubeFlutterCallKitPlugin.callController.startCall(handle: callId, videoEnabled: videoEnabled, uuid: callId)
+
+            SwiftConnectycubeFlutterCallKitPlugin.callController.answerCall(uuid: callId)
             result(true)
         }
         else if (call.method == "reportCallFinished"){