+++ lib/src/connectycube_flutter_call_kit.dart
@@ -168,12 +168,9 @@ class ConnectycubeFlutterCallKit {
 
   /// Report that the current active call has been accepted by your application
   ///
-  static Future<void> reportCallAccepted({
-    required String? sessionId,
-    required int? callType,
-  }) async {
-    return _methodChannel.invokeMethod(
-        "reportCallAccepted", {'session_id': sessionId, 'call_type': callType});
+  static Future<void> reportCallAccepted({required String? sessionId}) async {
+    return _methodChannel
+        .invokeMethod("reportCallAccepted", {'session_id': sessionId});
   }
 
   /// Report that the current active call has been ended by your application